chore: clean test filenames
