+++ test/context_test.mocks.dart
@@ -1,0 +1,88 @@
+// Mocks generated by Mockito 5.4.5 from annotations
+// in absmartly_sdk/test/context_test.dart.
+// Do not manually edit this file.
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i2;
+
+import 'package:absmartly_sdk/context.dart' as _i6;
+import 'package:absmartly_sdk/context_data_provider.dart' as _i3;
+import 'package:absmartly_sdk/context_event_handler.dart' as _i5;
+import 'package:absmartly_sdk/context_event_logger.dart' as _i8;
+import 'package:absmartly_sdk/json/context_data.dart' as _i4;
+import 'package:absmartly_sdk/json/publish_event.dart' as _i7;
+import 'package:mockito/mockito.dart' as _i1;
+
+// ignore_for_file: type=lint
+// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
+
+class _FakeCompleter_0<T> extends _i1.SmartFake implements _i2.Completer<T> {
+  _FakeCompleter_0(Object parent, Invocation parentInvocation)
+    : super(parent, parentInvocation);
+}
+
+/// A class which mocks [ContextDataProvider].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockContextDataProvider extends _i1.Mock
+    implements _i3.ContextDataProvider {
+  @override
+  _i2.Completer<_i4.ContextData> getContextData() =>
+      (super.noSuchMethod(
+            Invocation.method(#getContextData, []),
+            returnValue: _FakeCompleter_0<_i4.ContextData>(
+              this,
+              Invocation.method(#getContextData, []),
+            ),
+            returnValueForMissingStub: _FakeCompleter_0<_i4.ContextData>(
+              this,
+              Invocation.method(#getContextData, []),
+            ),
+          )
+          as _i2.Completer<_i4.ContextData>);
+}
+
+/// A class which mocks [ContextEventHandler].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockContextEventHandler extends _i1.Mock
+    implements _i5.ContextEventHandler {
+  @override
+  _i2.Completer<void> publish(_i6.Context? context, _i7.PublishEvent? event) =>
+      (super.noSuchMethod(
+            Invocation.method(#publish, [context, event]),
+            returnValue: _FakeCompleter_0<void>(
+              this,
+              Invocation.method(#publish, [context, event]),
+            ),
+            returnValueForMissingStub: _FakeCompleter_0<void>(
+              this,
+              Invocation.method(#publish, [context, event]),
+            ),
+          )
+          as _i2.Completer<void>);
+}
+
+/// A class which mocks [ContextEventLogger].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockContextEventLogger extends _i1.Mock
+    implements _i8.ContextEventLogger {
+  @override
+  void handleEvent(_i6.Context? context, _i8.EventType? type, dynamic data) =>
+      super.noSuchMethod(
+        Invocation.method(#handleEvent, [context, type, data]),
+        returnValueForMissingStub: null,
+      );
+}
